python/packages/liman_finops/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "liman_finops"
dynamic = ["version"]
description = "Liman AgentOps FinOps Library"
readme = "README.md"
authors = [
    { name = "Guro Bokum", email = "jiojiajiu@gmail.com" }
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "opentelemetry-api>=1.35.0",
    "opentelemetry-instrumentation>=0.56b0",
    "opentelemetry-sdk>=1.35.0",
    "wrapt>=1.17.2",
]

[tool.hatch.version]
path = "VERSION"
pattern = "(?P<version>.+)"

[tool.poe]
include = "../../shared_poe.toml"

[tool.poe.tasks]
mypy = "mypy --config-file ../../pyproject.toml ."

[tool.ruff]
extend = "../../pyproject.toml"
